various changes for icons
